[core-word-engine] Profanity Checking and Score Improvements
In order to prevent inappropriate submissions, we've included a profanity checker.
Additionally, we've improved are ability to check for english words via pyenchant.
Lastly, we've set our scoring to be proportional to the word length.
That is, for every new-char usage, we decrease the potential score by (1/wordlen) * num_new_char_uses

NVG.